Revert "更新框架"
This reverts commit 26bc3c38f0561ba6c21c5571fde6f55fcb91ae27.

+++ Gksyb.Common/Data/DbColumnInfo.cs
@@ -30,7 +30,7 @@
         /// <summary>
         /// 最大长度
         /// </summary>
-        public long? MaxLength { get; set; }
+        public int? MaxLength { get; set; }
 
         /// <summary>
         /// 整数长度
